Add error message if wget fails in powerswitch-set
Using powerswitch-set, if you don't have the right
password, or some other error occurs, then you don't
get any feedback on the problem.  This is partly due to
us using the wget -q switch.  In any event, check the wget
return code and show a message, so the failure is not silent.