Add teensy-usb.py program
Teensy is an arduino board with some nice USB capabilities.
If a teensy board is programmed with a sketch called "etherkey",
then it can be used for a host to control keystrokes sent to
another machine.

teensy-usb.py is a program that allows you to control from a host
the keystrokes to send to another machine (via the teensy board)
through the USB port of the other machine.