Have 'ttc run' use exec
When using ttc run, you want the command to output exactly what
comes from the target as a result of the run command's output.
This will be true for run_cmds using ssh or adb.  We do NOT
want to interpret the error code, or interfere with the data stream
(stdout or stderr) in any way.  Therefore, switch the run command
to use do_run() instead of exec_cmd().

exec_cmd buffers all output before printing it, and emits an
error when the rcode is non-zero (and changes the exit code).
All of this is undesirable when doing a 'ttc run' operation.

Execute the last command in the run_cmd command block
using os.execvp().