Fix quoting for ttc run
If the run_cmd value uses quotes around "$COMMAND" (like that),
then keep the cmd_to_run value together, instead of splitting
it into separate arguments.  This was necessary to prevent
'grabserial ... -c "$COMMAND' where $COMMAND was 'uname -a'
from separating the 'uname' and '-a' and having grabserial
interpret the '-a' as one of it's arguments.

This solution is a bit hacky, but maintaining correct quoting
is a thorny problem.  Every instance of $COMMAND I've seen
has either used $COMMAND (alone), or "$COMMAND" (quoted) only.
So this should work for now.