Convert to version 2.0 - finishing python 3 support
Convert remaining tabs to spaces.  Eliminate remaining uses
of print ..., (print ending with a comma).  Add VIM config line
to keep tabs from creeping in in the future.

And convert VERSION to 2.0.0 (and put python interpreter back
to unqualified (not specifically version 3)

Save the previous ttc version as ttc2.  This might be needed
for git blame.