Switch do_remote_command to .._exec
Execute remote commands using os.execvp() instead of getstatusouput.
This make execution of commands potentially interactive and
non-blocking.  It also might preserve argument units (but I'm still
investigating that).  Add 'login' back into the list of
easy_remote_commands.

Also, add '-t' as an argument to ssh invocation from my host to the
board's host.  This forces creation of a tty, which is needed for
interactivity between my host and the board.

Signed-off-by: Tim Bird <tim.bird@sony.com>