Add exec_cmd and cmd_args to host entry
Allow user to specify the command to use to execute remote
ttc operations on a host.  Add the 'exec_cmd' and 'cmd_args'
attributes to the host entry.  If these are not present, their
values default to '/usr/bin/ssh', and a long string of ssh
options (see the code).

Theoretically, this allows for customization of the ssh args,
or use of an entirely different command to perform remote ttc
operations (e.g. fsh).  Implementing an alternate exec_cmd
is left as an exercise for the reader.

Signed-off-by: Tim Bird <tim.bird@sony.com>