Add support for duration to 'ttc reserve'
Add needed fields to reservation file.  Create a reservation class,
and routines for checking reservation expiration and parsing
the duration time string.  Add a --no-check flag, that is used
when doing a command status operation (so that even if a user
doesn't have the board reserved, they can check on its status).

This flag is only intended for internal use (when ttc is calling
itself).

Refactor import of the time module, and add an import for the datetime
module.

Signed-off-by: Tim Bird <tim.bird@sony.com>