Suppress <host>:Targets as a result on 'ttc list'
When getting targets from a remote host, and the '-q' option
is not used, it returns the string "Targets on this host:" as
one of the lines.  Fix the filter string for this line so
it actually works (by removing the ':').  This will cause
this line to be suppressed and not interpreted as one of the
targets on that host.

Signed-off-by: Tim Bird <tim.bird@sony.com>