Don't use --debug in 'ttc run' inside command_status()
The --debug option causes the the nested 'ttc run' command to
output a version of the string that is used to test whether
the board can run commands.  This causes failed commands to
be reported as success (and the command status to be reported
as "OPERATIVE"), incorrectly.

Also add a check for non-zero rcode.