Add support for -v to 'ttc list' command
If '-v' is used with 'ttc list', show the value of the 'ttc_conf_file'
attribute, in the list of targets.  Set ttc_conf_file during reading of
the configs.  Document the -v option in the usage help for 'list'.

Also, fix a typo on a warning message for missing attribute.