Add warning about merging target configs
I was stumped when some attributes for one of my target weren't
reflected correctly in 'ttc info' for that target.  It turns out
I had a duplicate config for that target in a separate file.

Add some debug lines to help people see where configuration comes
from, as well as a warning when config attributes are merged from
two different files.